test(dynamic-forms): run internal entry-point specs and co-locate relocated addon specs
The vitest config only included src and integration spec globs, so every
spec under internal/src was silently dormant. Add the internal/src glob,
which activates roughly 1080 previously-unrun tests (2078 to 3162).

Eight of those specs had stale relative imports left over from the original
internal/main entry-point split (paths to testing/, state/FormStateManager,
the logger, and the registry services, plus a moved FORM_OPTIONS token).
Repoint them to the current locations.

Move the three addon specs whose source was relocated to /internal in this
PR alongside that source (df-addon-slot, inject-addon-type-registry,
run-preset-action), using relative imports per the internal no-barrel rule,
so the tests live with the code they cover.
