+++ packages/dynamic-forms/internal/src/lib/core/expressions/condition-evaluator.spec.ts
@@ -2,7 +2,7 @@ import { vi } from 'vitest';
 import { ConditionalExpression } from '../../models/expressions/conditional-expression';
 import { EvaluationContext } from '../../models/expressions/evaluation-context';
 import { evaluateCondition } from './condition-evaluator';
-import { createMockLogger, MockLogger } from '../../../../testing/src/mock-logger';
+import { createMockLogger, MockLogger } from '../../../../../testing/src/mock-logger';
 
 describe('condition-evaluator', () => {
   let mockContext: EvaluationContext;
