+++ packages/dynamic-forms/internal/src/lib/core/expressions/logic-function-factory.spec.ts
@@ -5,11 +5,13 @@ import { FieldContext } from '@angular/forms/signals';
 import { of } from 'rxjs';
 import { vi } from 'vitest';
 import { ConditionalExpression } from '../../models/expressions/conditional-expression';
-import { FunctionRegistryService, FieldContextRegistryService, RootFormRegistryService } from '../registry';
-import { FormStateManager } from '../../state/form-state-manager';
+import { FunctionRegistryService } from '../registry/function-registry.service';
+import { FieldContextRegistryService } from '../registry/field-context-registry.service';
+import { RootFormRegistryService } from '../registry/root-form-registry.service';
+import { FormStateManager } from '../../../../../src/lib/state/form-state-manager';
 import { createLogicFunction, createDebouncedLogicFunction } from './logic-function-factory';
 import { DynamicFormLogger } from '../../providers/features/logger/logger.token';
-import { createMockLogger, MockLogger } from '../../../../testing/src/mock-logger';
+import { createMockLogger, MockLogger } from '../../../../../testing/src/mock-logger';
 import { DEPRECATION_WARNING_TRACKER } from '../../utils/deprecation-warning-tracker';
 import { createWarningTracker } from '../../utils/warning-tracker';
 import { HTTP_CONDITION_CACHE, HttpConditionCache } from '../http/http-condition-cache';
