+++ packages/dynamic-forms/internal/src/lib/core/registry/function-registry.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { EvaluationContext } from '../../models/expressions/evaluation-context';
 import { FunctionRegistryService } from './function-registry.service';
-import { createMockLogger } from '../../../../testing/src/mock-logger';
+import { createMockLogger } from '../../../../../testing/src/mock-logger';
 import type { Logger } from '../../providers/features/logger/logger.interface';
 
 describe('FunctionRegistryService', () => {
