+++ packages/dynamic-forms/internal/src/lib/core/validation/validator-factory.spec.ts
@@ -3,10 +3,12 @@ import { Injector, runInInjectionContext, signal } from '@angular/core';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { form, schema } from '@angular/forms/signals';
 import { ValidatorConfig } from '../../models/validation/validator-config';
-import { RootFormRegistryService, FunctionRegistryService, FieldContextRegistryService } from '../registry';
-import { FormStateManager } from '../../state/form-state-manager';
-import { DynamicFormLogger } from '../registry/field-context-registry.service';
-import { DEPRECATION_WARNING_TRACKER } from '../registry/field-context-registry.service';
+import { RootFormRegistryService } from '../registry/root-form-registry.service';
+import { FunctionRegistryService } from '../registry/function-registry.service';
+import { FieldContextRegistryService } from '../registry/field-context-registry.service';
+import { FormStateManager } from '../../../../../src/lib/state/form-state-manager';
+import { DynamicFormLogger } from '../../providers/features/logger/logger.token';
+import { DEPRECATION_WARNING_TRACKER } from '../../utils/deprecation-warning-tracker';
 import { createWarningTracker } from '../../utils/warning-tracker';
 import { applyValidator, applyValidators } from './validator-factory';
 import { LogicFunctionCacheService } from '../expressions/logic-function-cache.service';
