+++ packages/dynamic-forms/internal/src/lib/core/values/dynamic-value-factory.spec.ts
@@ -3,10 +3,11 @@ import { FieldContext, FieldTree } from '@angular/forms/signals';
 import { signal, Injector, runInInjectionContext } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { createDynamicValueFunction } from './dynamic-value-factory';
-import { RootFormRegistryService, FieldContextRegistryService } from '../registry';
-import { FormStateManager } from '../../state/form-state-manager';
-import { DynamicFormLogger } from '../registry/field-context-registry.service';
-import { ConsoleLogger } from '../../providers/features/logger/console-logger';
+import { RootFormRegistryService } from '../registry/root-form-registry.service';
+import { FieldContextRegistryService } from '../registry/field-context-registry.service';
+import { FormStateManager } from '../../../../../src/lib/state/form-state-manager';
+import { DynamicFormLogger } from '../../providers/features/logger/logger.token';
+import { ConsoleLogger } from '../../../../../src/lib/providers/features/logger/console-logger';
 import { DynamicValueFunctionCacheService } from './dynamic-value-function-cache.service';
 
 describe('dynamic-value-factory', () => {
