+++ packages/dynamic-forms/vite.config.ts
@@ -52,7 +52,11 @@ export default defineConfig(({ mode }) => {
       hookTimeout: 1000,
       // Retry flaky tests due to browser mode module mocking inconsistencies
       retry: 2,
-      include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}', 'integration/src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
+      include: [
+        'src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}',
+        'integration/src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}',
+        'internal/src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}',
+      ],
       reporters: ['default'],
       coverage: {
         enabled: true,
